refactor: update import paths for UI components to improve consistency
