+++ frontend/components/ui/item_misc_fields_ui.py
@@ -1,7 +1,7 @@
 from typing import Optional, Tuple
 
 import streamlit as st
-from components.ui.filter_utils import WIDGET_KEY_PREFIX
+from components.ui.item_filter_utils import WIDGET_KEY_PREFIX
 
 
 def render_misc_fields() -> Tuple[Optional[str], Optional[str], Optional[str]]:
